client/backend/guest_management/face-service/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
facenet-pytorch==2.6.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=11.0.0
python-multipart==0.0.17
numpy>=1.24.0
